@@ -25,7 +25,9 @@
 #include <io.h>
 #include <objbase.h>  // guid stuff
 #include <shellapi.h>
+#include <ShlObj.h>
 #include <windows.h>
+#include <winerror.h>
 #else
 #include <dirent.h>
 #include <errno.h>
@@ -755,19 +757,49 @@ std::string& GetExeDirectory()
 	return DolphinPath;
 }
 
+std::string GetHomeDirectory()
+{
+	std::string homeDir;
+#ifdef _WIN32
+	wchar_t *path = nullptr;
+	
+	if (SUCCEEDED(SHGetKnownFolderPath(FOLDERID_Documents, 0, nullptr, &path))) {
+		char pathStr[MAX_PATH];
+		wcstombs(pathStr, path, MAX_PATH);
+	
+		homeDir = std::string(pathStr);
+		CoTaskMemFree(path);		
+	}
+	else {
+		const char* home = getenv("USERPROFILE");
+		homeDir = std::string(home) + "\\Documents";
+	}
+#else
+	const char* home = getenv("HOME");
+    homeDir = std::string(home);
+#endif
+
+	return homeDir;
+}
+
 std::string GetSysDirectory()
 {
 	std::string sysDir;
 
 #if defined(__APPLE__)
-	sysDir = GetBundleDirectory() + DIR_SEP + SYSDATA_DIR;
+	sysDir = GetBundleDirectory() + DIR_SEP + SYSDATA_DIR + DIR_SEP;
 #elif defined(_WIN32) || defined(LINUX_LOCAL_DEV)
-	sysDir = GetExeDirectory() + DIR_SEP + SYSDATA_DIR;
+	sysDir = GetExeDirectory() + DIR_SEP + SYSDATA_DIR + DIR_SEP;
 #else
-	sysDir = SYSDATA_DIR;
+	const char* home = getenv("HOME");
+	if (!home) home = getenv("PWD");
+	if (!home) home = "";
+	std::string home_path = std::string(home) + DIR_SEP;
+	const char* config_home = getenv("XDG_CONFIG_HOME");
+	sysDir = std::string(config_home && config_home[0] == '/' 
+		? config_home : (home_path + ".config")) 
+		+ DIR_SEP DOLPHIN_DATA_DIR DIR_SEP "Sys" DIR_SEP;
 #endif
-	sysDir += DIR_SEP;
-
 	INFO_LOG(COMMON, "GetSysDirectory: Setting to %s:", sysDir.c_str());
 	return sysDir;
 }
@@ -971,7 +1003,7 @@ bool IOFile::OpenShared(const std::string& filename, const char openmode[], int 
 {
 	Close();
 #ifdef _WIN32
-	m_file = _fsopen(filename.c_str(), openmode, shflag);
+	m_file = _tfsopen(UTF8ToTStr(filename).c_str(), UTF8ToTStr(openmode).c_str(), shflag);
 #else
 	m_file = fopen(filename.c_str(), openmode);
 #endif
