@@ -23,7 +23,7 @@
 //#else
 //" " BUILD_TYPE_STR " " SCM_DESC_STR;
 //#endif
-#define SLIPPI_REV_STR "2.0.1"
+#define SLIPPI_REV_STR "2.1.1"
 const std::string scm_rev_str = "Faster Melee - Slippi (" SLIPPI_REV_STR ")";
 const std::string scm_slippi_semver_str = SLIPPI_REV_STR;
 
