@@ -49,8 +49,11 @@ class SlippiMatchmaking
 	std::string GetErrorMessage();
 
   protected:
-	const std::string MM_HOST = "35.247.98.48";
+	const std::string MM_HOST_DEV = "35.197.121.196"; // Dev host
+	const std::string MM_HOST_PROD = "35.247.98.48";  // Production host
 	const u16 MM_PORT = 43113;
+
+	std::string MM_HOST = "";
 
 	ENetHost *m_client;
 	ENetPeer *m_server;
