@@ -81,8 +81,12 @@ static void Read()
 	int payload_size = 0;
 	while (s_adapter_thread_running.IsSet())
 	{
-		adapter_error = libusb_interrupt_transfer(s_handle, s_endpoint_in, s_controller_payload_swap,
-			sizeof(s_controller_payload_swap), &payload_size, 16) != LIBUSB_SUCCESS && SConfig::GetInstance().bAdapterWarning;
+		int err = libusb_interrupt_transfer(s_handle, s_endpoint_in, s_controller_payload_swap,
+		                                    sizeof(s_controller_payload_swap), &payload_size, 16);
+		adapter_error = err != LIBUSB_SUCCESS && SConfig::GetInstance().bAdapterWarning;
+
+		if (err)
+			ERROR_LOG(SERIALINTERFACE, "adapter libusb read failed: err=%s", libusb_error_name(err));
 
 		{
 			std::lock_guard<std::mutex> lk(s_mutex);
@@ -100,7 +104,12 @@ static void Write()
 	while (s_adapter_thread_running.IsSet())
 	{
 		if (s_rumble_data_available.WaitFor(std::chrono::milliseconds(100)))
-			libusb_interrupt_transfer(s_handle, s_endpoint_out, s_latest_rumble_data, sizeof(s_latest_rumble_data), &size, 16);
+		{
+			int err = libusb_interrupt_transfer(s_handle, s_endpoint_out, s_latest_rumble_data,
+			                                    sizeof(s_latest_rumble_data), &size, 16);
+			if (err)
+				ERROR_LOG(SERIALINTERFACE, "adapter libusb write failed: err=%s", libusb_error_name(err));
+		}
 	}
 
 	s_rumble_data_available.Reset();
